Split out extension of BuildConfiguratio
