+++ Sources/ScipioKit/Producer/PIF/BuildConfiguration+Extensions.swift
@@ -1,0 +1,12 @@
+import XCBuildSupport
+
+extension PIF.BuildConfiguration {
+    mutating func setImpartedBuildProperties(_ newValue: PIF.ImpartedBuildProperties) {
+        self = PIF.BuildConfiguration(
+            guid: guid,
+            name: name,
+            buildSettings: buildSettings,
+            impartedBuildProperties: newValue
+        )
+    }
+}
