+++ Sources/ScipioKit/URL+Extensions.swift
